requirements.txt
fastapi
uvicorn
openai
python-dotenv
requests
psutil
pyautogui
pydantic
sentence-transformers
llama-index
spacy
easyocr
PyPDF2  # Добавили!
python-docx
pdfminer.six==20221105
chromadb>=0.5,<0.6